Move AbstractFormDumper to subpackage to ensure proper class resolution during gradle buildModel
The gradle build tool only replaces/build different classes during the build process if they are explicitly imported. By moving the AbstractFormDumper to a subpackage, we ensure that it is correctly imported and resolved at runtime.

Ticket: SBW-33072
