+++ scripts/commons/serviceportal/forms/formdumper/CsvDumper.groovy
@@ -1,5 +1,6 @@
 package commons.serviceportal.forms.formdumper
 
+import commons.serviceportal.forms.formdumper.dummy.AbstractFormDumper
 import de.seitenbau.serviceportal.scripting.api.v1.ScriptingApiV1
 import de.seitenbau.serviceportal.scripting.api.v1.form.FieldGroupInstanceV1
 import de.seitenbau.serviceportal.scripting.api.v1.form.FieldTypeV1
@@ -7,14 +8,12 @@ import de.seitenbau.serviceportal.scripting.api.v1.form.FormFieldV1
 import de.seitenbau.serviceportal.scripting.api.v1.form.content.BinaryContentV1
 import de.seitenbau.serviceportal.scripting.api.v1.form.content.FormContentV1
 
-import java.text.SimpleDateFormat
-
 /**
  * Transforms a Serviceportal-proprietary form (= a FormContentV1 object) into CSV files.
  * Fields are seperated by a separator character (see constructor) and always surrounded by quotes.
  * Individual fields are listed in rows, not columns. I.e. the output is "tall" (instead of "wide").
  *
- * See {@link AbstractFormDumper}
+ * See {@link commons.serviceportal.forms.formdumper.dummy.AbstractFormDumper}
  *
  * <h2>Example Output:</h2>
  *
