+++ scripts/commons/serviceportal/forms/formdumper/TextDumper.groovy
@@ -1,5 +1,6 @@
 package commons.serviceportal.forms.formdumper
 
+import commons.serviceportal.forms.formdumper.dummy.AbstractFormDumper
 import de.seitenbau.serviceportal.scripting.api.v1.ScriptingApiV1
 import de.seitenbau.serviceportal.scripting.api.v1.form.FieldGroupInstanceV1
 import de.seitenbau.serviceportal.scripting.api.v1.form.FormFieldV1
@@ -8,7 +9,7 @@ import de.seitenbau.serviceportal.scripting.api.v1.form.content.FormContentV1
 /**
  * Transforms a Serviceportal-proprietary form (= a FormContentV1 object) into raw, human-readable text.
  *
- * See {@link AbstractFormDumper}
+ * See {@link commons.serviceportal.forms.formdumper.dummy.AbstractFormDumper}
  *
  * <h2>Example Output:</h2>
  *
